+++ media_driver/agnostic/common/codec/hal/codechal_decoder.h
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2011-2018, Intel Corporation
+* Copyright (c) 2011-2019, Intel Corporation
 *
 * Permission is hereby granted, free of charge, to any person obtaining a
 * copy of this software and associated documentation files (the "Software"),
@@ -902,10 +902,10 @@ private:
         uint32_t allocHeight,
         MOS_FORMAT format);
     //!
-    //! \brief    Resize specific reference surfaces 
+    //! \brief    Resize specific reference surfaces
     //! \details  Resize specific reference surfaces for decode downsampling for all codec types
     //! \param    frameIdx
-    //!           [in] index of surfaces array 
+    //!           [in] index of surfaces array
     //! \param    width
     //!           [in] Width of the surfaces to be allocated
     //! \param    height
@@ -921,7 +921,7 @@ private:
         uint32_t     height,
         MOS_FORMAT   format);
     //!
-    //! \brief    Deallocate specific reference surfaces 
+    //! \brief    Deallocate specific reference surfaces
     //! \details  Deallocate specific reference surfaces for decode downsampling for all codec types
     //! \param    frameIdx
     //!           [in] index of surfaces array
@@ -1046,9 +1046,9 @@ protected:
     bool                        m_disableLockForTranscode = false;
 
     //! \brief Indicate how many passes is needed to finish decoding a picture
-    //! \details Initialize decode pass number to 1, for those decoder need more than 1 decoding pass, 
+    //! \details Initialize decode pass number to 1, for those decoder need more than 1 decoding pass,
     //!          modify this value in specific decoder files.
-    uint8_t                     m_decodePassNum     = 1;
+    uint16_t                     m_decodePassNum     = 1;
 
     //! \brief MMIO Hcp Frame CRC report offset
     uint32_t                    m_hcpFrameCrcRegOffset = 0;
